fixing code review issues
